pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hapi"
version = "0.1.0"
description = "Host and VPS discovery/control plane service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "pyyaml>=6.0.1",
  "fastmcp>=1.12.0",
  "httpx>=0.27.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-mock>=3.14.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
